Add basic (unchecked) authentication
The server and client have been split into smaller functions, handling
reading or writing and authentication handshaking. Additionally, we
don't thread log around as much in server as we don't need it everywhere
we `map_err` -- we only need it on the final `map_err`.

The client now prints out all understood messages from the server and,
after a 5s delay, sends a Goodbye and exists.