Split auth messages out of server/client message kinds
The server and client will only exchange the handshake messages at the
beginning of the connection. Having them in the same type as regular
messages requires us to match against them when reciving messages when
we know we won't get them. To that end, they have been separated out
into their own types.

The `Send` and `Recv` type aliases are no longer generic over the
underlying stream. They are now generic over the message kinds they
send/receive and are concrete over the stream.

Now that we have multiple message kinds, it is useful to convert the
underlying streams into `Send` and `Recv` for the required kind. We now
have `split_stream`, which takes a `TcpStream` and turns it into a
`Send` and a `Recv`, and `join_stream`, which does the reverse. This
allows us to easily use different message types in different parts of
the client/server lifetime.